+++ src/data/string-validation.ts
